% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignmentQA.R
\name{assignmentQA}
\alias{assignmentQA}
\title{Sum formula assignment QA}
\usage{
assignmentQA(assignment_output, junction_output)
}
\arguments{
\item{assignment_output}{The result of the assignSumFormulas function}

\item{junction_output}{The result of the junction function}
}
\value{
A data frame containing the intensity ratios of sum formulas relative to the raw data.
}
\description{
Computes the ratio of the summed intensity of peaks with sum formula assignment to the raw data
}
