% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignSumFormulas.R
\name{assignSumFormulas}
\alias{assignSumFormulas}
\title{Sum Formula assignment}
\usage{
assignSumFormulas(
  junction_output,
  ionizationMode,
  LowMassLimit = 100,
  HighMassLimit = 1000,
  assign_error = 0.7,
  set_Nx = 4,
  set_Sx = 1,
  set_Px = 0,
  set_S34x = 0,
  set_N15x = 0,
  set_Dx = 0,
  set_Ex = 0,
  set_Clx = 0,
  set_Cl37x = 0,
  set_Fx = 0,
  set_Brx = 0,
  set_Br81x = 0,
  set_Ix = 0,
  set_Mx = 0,
  set_NH4x = 0,
  set_Zx = 1,
  set_Sval = 2,
  set_Nval = 3,
  set_S34val = 2,
  set_N15val = 3,
  set_Pval = 5,
  set_Ox = 30,
  set_O_Cmin = 0,
  set_O_Cmax = 2.5,
  set_H_Cmin = 0.3,
  set_H_Cmax = 3,
  set_DBEOmin = -13,
  set_DBEOmax = 13,
  set_Omin = 0,
  set_POEx = 0,
  set_NOEx = 0,
  set_max_def = 0.9,
  set_min_def = 0.5,
  set_HetCut = "off",
  set_NMScut = "on",
  set_DeNovo = 300,
  set_nLoop = 5,
  set_SulfCheck = "on",
  set_Ambig = "off",
  set_MSMS = "off",
  set_S34_abund = 30,
  set_C13_abund = 60,
  set_N3corr = "on"
)
}
\arguments{
\item{junction_output}{The result of the junction function}

\item{ionizationMode}{"neg" or "pos"}

\item{LowMassLimit}{Lower end of the measured mass range in Da (default 100)}

\item{HighMassLimit}{Upper end of the measured mass range in Da (default 1000)}

\item{assign_error}{Error tolerance for sum formula assignment in ppm (default 0.7)}

\item{set_Nx}{Number of Nitrogen atoms allowed (default 4)}

\item{set_Sx}{Number of Sulphur atoms allowed (default 1)}

\item{set_Px}{Number of Phosphorous atoms allowed (default 0)}

\item{set_S34x}{Number of Sulphur-34 atoms allowed (default 0)}

\item{set_N15x}{Number of Nitrogen-15 atoms allowed (default 0)}

\item{set_Dx}{Number of Deuterium atoms allowed (default 0)}

\item{set_Ex}{Number of Carbon-13 atoms allowed (default 0)}

\item{set_Clx}{Number of Chlorine atoms allowed (default 0)}

\item{set_Cl37x}{Number of Chlorine-37 atoms allowed (default 0)}

\item{set_Brx}{Number of Bromine atoms allowed (default 0)}

\item{set_Br81x}{Number of Bromine-81 atoms allowed (default 0)}

\item{set_Ix}{Number of Iodine atoms allowed (default 0)}

\item{set_Mx}{Number of Sodium adducts allowed (default 0)}

\item{set_NH4x}{Number of Ammonium adducts allowed (default 0)}

\item{set_Zx}{Amount of charge allowed (default 1)}
}
\value{
A data frame containing the assigned sum formulas and their intensities per sample
}
\description{
Assigns sum formulas to m/z values using MFAssignR. Refer to MFAssignR::MFAssign for detailed instructions and help.
}
