% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excludeRecalErrors.R
\name{excludeRecalErrors}
\alias{excludeRecalErrors}
\title{Exclude recalibration errors}
\usage{
excludeRecalErrors(recal_result)
}
\arguments{
\item{recal_result}{The result of the recalibrate function}
}
\value{
A list containing for each sample the recalibrated m/z values, the noise level, an indicator of recailbration success or failure, and the sample name.
}
\description{
Removes spectra that couldn't be recalibrated from the dataset
}
