% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importMzml.R
\name{importMzml}
\alias{importMzml}
\title{mzml file import}
\usage{
importMzml(filePath, SN, parallel = TRUE)
}
\arguments{
\item{filePath}{The path to the folder containing the .mzml files}

\item{SN}{Signal to noise ratio (noise level will be determined automatically)}

\item{parallel}{Parallel processing (TRUE or FALSE, default: TRUE)}
}
\value{
A list containing for each sample the spectra, noise level, and sample name
}
\description{
Read centroided spectra from .mzml files
}
