% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction.R
\name{junction}
\alias{junction}
\title{Junction of m/z tables}
\usage{
junction(recal_result, tolerance = 0.5)
}
\arguments{
\item{recal_result}{The result of the recalibrate function}

\item{tolerance}{Tolerance in joining close m/z values (in ppm, default 0.5)}
}
\value{
A dataframe containing one column of m/z values followed by their respective intensities in all spectra.
}
\description{
Joins together the m/z tables of all recalibrated mass spectra.
}
