% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalEval.R
\name{recalEval}
\alias{recalEval}
\title{Evaluate recalibrations}
\usage{
recalEval(import, recal_result)
}
\arguments{
\item{import}{The result of the import Mzml function}

\item{recal_result}{The result of the recalibrate function}
}
\value{
Printout of an overview of the recalibration success or failure for each sample.
}
\description{
Prints the recalibration outcome for all samples.
}
