% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalibrate.R
\name{recalilbrate}
\alias{recalilbrate}
\title{Mass spectra recalibration}
\usage{
recalilbrate(
  import,
  ionizationMode,
  LowMassLimit,
  HighMassLimit,
  CHO_assign_error = 2.5,
  SN_recal = 10,
  RecalRangeCheck = TRUE,
  recalRange = c(150, 500),
  AutoUpperLimit = TRUE,
  AutoUpperLimitTarget = 0.8,
  Nitrogen_recal = 0,
  max_loop = 3,
  AttemptFixingErrors = FALSE,
  parallel = TRUE
)
}
\arguments{
\item{import}{The result of the importMzml function}

\item{ionizationMode}{"neg" or "pos"}

\item{LowMassLimit}{Lower end of the measured mass range in Da (default 100)}

\item{HighMassLimit}{Upper end of the measured mass range in Da (default 1000)}

\item{CHO_assign_error}{Error tolerance for preliminary sum formula assignment in ppm (default 2.5)}

\item{SN_recal}{Minimum signal-to-noise ratio for peaks to be used as potential recalibrants (default 10)}

\item{RecalRangeCheck}{TRUE/FALSE  Require that the recalibration covers a specified mass range (set below, default TRUE)}

\item{recalRange}{TRUE/FALSE  Sample must be recalibrated within this entire range, else the sample is reported as having failed recalibration (default c(50, 500))}

\item{AutoUpperLimit}{TRUE/FALSE Should the upper limit of recalRange be set automatically to the m/z where a set percentage of the cumulative intensity is reached? (default TRUE)}

\item{AutoUpperLimitTarget}{Target percentage of cumulative intensity starting from the lowest detected m/z (default 0.8)}

\item{Nitrogen_recal}{Number of N atoms allowed in inital sum formula assignment (default 0)}

\item{max_loop}{Maximum number of attempts at fixing recalibration errors (default 3)}

\item{AttemptFixingErrors}{TRUE/FALSE In case recaclibration fails, try with just a single recalibration series instead. This setting automatically results in no RecalRangeCheck for those attempts. (default FALSE)}

\item{parallel}{Allow parallel execution? (default TRUE)}
}
\value{
A list containing for each sample the recalibrated m/z values, the noise level, an indicator of recailbration success or failure, and the sample name.
}
\description{
Recalibrates the input mass spectra using function MFAssignR::Recal. Automatically selects homologous series for each sample based on function MFAssignR::RecalList.
}
