% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalPlot.R
\name{recalPlot}
\alias{recalPlot}
\title{Recalibration plots}
\usage{
recalPlot(recal_result, sample_nr)
}
\arguments{
\item{recal_result}{The result of the recalibrate function}

\item{sample_nr}{The index of the spectrum to be plotted.}
}
\value{
A plot of a mass spectrum and recalibration errors.
}
\description{
Plots a recalibrated mass spectrum, highlighting the peaks used as recalibrants and their respective mass error in ppm.
}
