% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalRange.R
\name{recalRange}
\alias{recalRange}
\title{Recalibration ranges}
\usage{
recalRange(recal_result)
}
\arguments{
\item{recal_result}{The result of the recalibrate function}
}
\value{
A dataframe.
}
\description{
Displays the m/z of the lightest and heaviest masses used as recalibrants.
}
